Update fanctrl.py
The goal is to allow an automatic switch of the fan control strategy depending on whether the Laptop battery is discharging or not. 
The user should have the option to declare a separate strategy during discharge inside the config file. If none is given, no switch should happen.
If a separate strategy is given the following things have to be implemented:
- check regularly the charging state of the battery = done
- if the charging state changed, and the user has determined a separate fan control strategy, its values should be applied = to do