PR form new dedicated upstream branch, updated code per review comments (#5)
* Update fanctrl.py

The goal is to allow an automatic switch of the fan control strategy depending on whether the Laptop battery is discharging or not. 
The user should have the option to declare a separate strategy during discharge inside the config file. If none is given, no switch should happen.
If a separate strategy is given the following things have to be implemented:
- check regularly the charging state of the battery = done
- if the charging state changed, and the user has determined a separate fan control strategy, its values should be applied = to do

* Update fanctrl.py

* Update config.json

* auto switch fan profiles depending on charging status

This updated version enables the automatic switch between two different fan curves depending on whether the Laptop is charging or discharging.
There are two new functions: updateStrategy and getBatteryChargingStatus. UpdateStrategy is called from once from __init__ and regularly from run(). UpdateStrategy then calls getBatteryChargingStatus. getBatteryChargingStatus checks whether the charging status changed, compared to the last look up. It returns either no change, or the new charging status. updateStrategy only loads a new fan curve if the charging status changed.
The config.json file is only read once, during __init__. The config.json file contains a new key "strategyOnDischarging". The user is supposed to enter one of the defined fan curves which should be active during discharge. If no separate discharge strategy is specified, the tool uses the same strategy as for charging.

* new entry: strategyOnDischarging

The user has the option to specify a separate fan curve/strategy which is only active during discharge.

* Update README.md

* Update README.md

* Update README.md

* updated intro to give credit

* Update README.md

* battery charging status path 

The battery charging status path has now a default value, which can be overwritten by providing a different path in the config file.
The motivation behind this change is to allow for different file systems.

* battery charging status path

New key: batteryChargingStatusPath. Any entry overwrites the default look up path.

* battery charging status path

* typo

* Update README.md

* removed comment about fork

* default strategy remains lazy

* restructured code 

Restructured code so that `updateStrategy` is only called if there is a separate strategy for discharging specified inside the config file.

* added entry

added entry batteryChargingStatusPath

* Update README.md

* removed battery status form print

* explained separate default/discharging strategies