implement critical temperature
The critical temperature is a setting that can be set per-strategy
("criticalTemp") that sets a point at which the moving average is
ignored and only the current temperature is used. This is done to allow
having a large moving average interval without the worry that the fans
will not kick in if there is a sudden spike in temperature. If the
critical temperature is unset or set to null (not the string "null", but
regular null in JSON), then this setting will be ignored (as if it were
set to infinity).